<feat> 사진 저장, 삭제 메서드 정리 - existImageInFile 메서드를 통해 파일에 id 이미지가 있는지 확인 - 없다면 저장, 있다면 저장 실패 알림 - Delegate 패턴으로 전달

+++ Picterest/ViewModel/SavedViewModel.swift
@@ -32,9 +32,26 @@ final class SavedViewModel {
     
     func deletePhotoEntity(index: Int) {
         let photoEntity = photoEntities[index]
-        ImageFileManager.shared.deleteImage(id: photoEntity.id ?? "")
-        CoreDataManager.shared.deletePhotoEntity(photoEntity: photoEntity) {
-            self.photoEntities.remove(at: index)
+        guard let id = photoEntity.id else {
+            return
         }
+        
+        ImageFileManager.shared.existImageInFile(id: id) { exist in
+            if exist {
+                ImageFileManager.shared.deleteImage(id: photoEntity.id ?? "")
+                CoreDataManager.shared.deletePhotoEntity(photoEntity: photoEntity) { success in
+                    if success {
+                        self.photoEntities.remove(at: index)
+                    }
+                }
+            }
+        }
+        
+//        ImageFileManager.shared.deleteImage(id: photoEntity.id ?? "")
+//        CoreDataManager.shared.deletePhotoEntity(photoEntity: photoEntity) { success in
+//            if success {
+//                self.photoEntities.remove(at: index)
+//            }
+//        }
     }
 }
